// server/src/shared/constants.d.ts
export declare const WORLD_SIZE: {
    RADIUS: number;
    CENTER_X: number;
    CENTER_Y: number;
};
export declare const GAME_SETTINGS: {
    SNAKE_START_SPEED: number;
    SNAKE_BOOST_MULTIPLIER: number;
    SNAKE_TURN_SPEED: number;
    SNAKE_START_WIDTH: number;
    MAX_BOTS: number;
    DURATIONS: {
        [key: string]: number;
    };
};
//# sourceMappingURL=constants.d.ts.map